refactor(input): restructure input/ into keyboard/ and pointer/ submodules
Splits the flat input/ module into a cohesive hierarchy:

  input/
  ├── ctx.rs          — InputCtx: bundles mod_state, pointer_state, backend,
  │                     config_path and wayland_display into one struct,
  │                     eliminating the repeated 5-parameter list across all
  │                     handler call sites
  ├── events.rs       — BackendInputEventData enum + thin dispatcher
  ├── keyboard/
  │   ├── mod.rs      — handle_keyboard_input entry point
  │   ├── bindings.rs — binding lookup, action dispatch, repeat policy
  │   │                 (was key_actions.rs; duplicate compositor_binding_action
  │   │                  fns collapsed into one)
  │   ├── modkeys.rs  — XKB helpers, ModState mutation, modifier matching
  │   │                 (was utils.rs; is_modifier_keycode moved here from
  │   │                  input_events.rs where it didn't belong)
  │   └── spawn.rs    — child-process spawning for launch bindings
  │                     (extracted from key_actions.rs; SRP)
  └── pointer/
      ├── mod.rs      — re-exports shared types
      ├── axis.rs     — scroll/wheel handling (extracted from input_events.rs)
      ├── button.rs   — button press/release (was pointer_button.rs)
      ├── focus.rs    — surface hit-testing (was pointer_focus.rs)
      ├── motion.rs   — motion and drag (was pointer_motion.rs)
      └── resize.rs   — resize geometry helpers (was resize_helpers.rs)

Update call sites in backend/winit.rs and backend/tty/mod.rs to construct
InputCtx at the event dispatch boundary rather than threading five individual
arguments through every handler.

Public API surface (BackendInputEventData, handle_backend_input_event,
spawn_command, active_node_screen_rect, active_node_surface_transform_screen_details,
active_resize_geometry_screen) is unchanged.
