+++ crates/halley-wl/src/compositor/mod.rs
@@ -1,0 +1,1 @@
+pub mod surface_ops;
