+++ crates/halley-wl/src/input/mod.rs
@@ -1,23 +1,11 @@
-mod input_events;
-mod input_keyboard;
-mod input_pointer;
-mod key_actions;
-mod pointer_button;
-mod pointer_core;
-mod pointer_dispatch;
-mod pointer_drag;
-mod pointer_focus;
-mod pointer_frame;
-mod pointer_motion;
-mod pointer_motion_drag;
-mod pointer_motion_resize;
-mod pointer_resize;
-mod resize_helpers;
-mod utils;
+pub(crate) mod ctx;
+pub(crate) mod events;
+pub(crate) mod keyboard;
+pub(crate) mod pointer;
 
-pub(crate) use input_events::{BackendInputEventData, handle_backend_input_event};
-pub(crate) use key_actions::spawn_command;
-pub(crate) use resize_helpers::{
+pub(crate) use events::{BackendInputEventData, handle_backend_input_event};
+pub(crate) use keyboard::spawn::spawn_command;
+pub(crate) use pointer::{
     active_node_screen_rect, active_node_surface_transform_screen_details,
     active_resize_geometry_screen,
 };
