+++ crates/halley-wl/src/input/pointer/mod.rs
@@ -1,0 +1,10 @@
+pub(crate) mod axis;
+pub(crate) mod button;
+pub(crate) mod focus;
+pub(crate) mod motion;
+pub(crate) mod resize;
+
+pub(crate) use resize::{
+    active_node_screen_rect, active_node_surface_transform_screen_details,
+    active_resize_geometry_screen,
+};
