--- crates/halley-wl/src/input/pointer_motion_drag.rs
@@ -1,406 +1,0 @@
-use std::time::Instant;
-
-use crate::backend::interface::BackendView;
-use crate::interaction::types::{DragAxisMode, HitNode, ModState, PointerState};
-use crate::state::{ActiveDragState, Halley};
-
-use super::pointer_drag::begin_drag;
-use super::pointer_frame::ButtonFrame;
-
-pub(super) fn cluster_join_dwell_ms(st: &Halley) -> u64 {
-    st.runtime.tuning.cluster_dwell_ms
-}
-
-pub(super) fn update_cluster_join_candidate(
-    st: &mut Halley,
-    node_id: halley_core::field::NodeId,
-    monitor: &str,
-    desired_center: halley_core::field::Vec2,
-    now: Instant,
-) -> bool {
-    if st
-        .model
-        .field
-        .cluster_id_for_member_public(node_id)
-        .is_some()
-    {
-        st.input.interaction_state.cluster_join_candidate = None;
-        return false;
-    }
-    let Some(node) = st.model.field.node(node_id) else {
-        st.input.interaction_state.cluster_join_candidate = None;
-        return false;
-    };
-    if node.kind != halley_core::field::NodeKind::Surface {
-        st.input.interaction_state.cluster_join_candidate = None;
-        return false;
-    }
-
-    let mover_ext = if matches!(
-        node.state,
-        halley_core::field::NodeState::Active | halley_core::field::NodeState::Drifting
-    ) {
-        st.surface_window_collision_extents(node)
-    } else {
-        st.collision_extents_for_node(node)
-    };
-    let candidate = st.cluster_bloom_for_monitor(monitor).and_then(|open_cid| {
-        let cluster = st.model.field.cluster(open_cid)?;
-        if !cluster.is_collapsed() {
-            return None;
-        }
-        let core_id = cluster.core?;
-        let core = st.model.field.node(core_id)?;
-        let core_monitor = st
-            .model
-            .monitor_state
-            .node_monitor
-            .get(&core_id)
-            .map(String::as_str)
-            .unwrap_or(monitor);
-        if core_monitor != monitor {
-            return None;
-        }
-        let core_ext = st.collision_extents_for_node(core);
-        let gap = st.non_overlap_gap_world();
-        let mover_left = desired_center.x - mover_ext.left;
-        let mover_right = desired_center.x + mover_ext.right;
-        let mover_top = desired_center.y - mover_ext.top;
-        let mover_bottom = desired_center.y + mover_ext.bottom;
-        let core_left = core.pos.x - core_ext.left - gap;
-        let core_right = core.pos.x + core_ext.right + gap;
-        let core_top = core.pos.y - core_ext.top - gap;
-        let core_bottom = core.pos.y + core_ext.bottom + gap;
-        let touching_gap = mover_right >= core_left
-            && mover_left <= core_right
-            && mover_bottom >= core_top
-            && mover_top <= core_bottom;
-        touching_gap.then_some(open_cid)
-    });
-
-    let Some(cluster_id) = candidate else {
-        st.input.interaction_state.cluster_join_candidate = None;
-        return false;
-    };
-    let now_ms = st.now_ms(now);
-    let keep_started_at = st
-        .input
-        .interaction_state
-        .cluster_join_candidate
-        .as_ref()
-        .filter(|existing| {
-            existing.cluster_id == cluster_id
-                && existing.node_id == node_id
-                && existing.monitor == monitor
-        })
-        .map(|existing| existing.started_at_ms)
-        .unwrap_or(now_ms);
-    let dwell_ms = cluster_join_dwell_ms(st);
-    st.input.interaction_state.cluster_join_candidate = Some(crate::state::ClusterJoinCandidate {
-        cluster_id,
-        node_id,
-        monitor: monitor.to_string(),
-        started_at_ms: keep_started_at,
-        ready: now_ms.saturating_sub(keep_started_at) >= dwell_ms,
-    });
-    false
-}
-
-pub(super) fn maybe_begin_core_drag_from_pending_press(
-    st: &mut Halley,
-    ps: &mut PointerState,
-    backend: &impl BackendView,
-    local_w: i32,
-    local_h: i32,
-    effective_sx: f32,
-    effective_sy: f32,
-    local_sx: f32,
-    local_sy: f32,
-    pointer_world: halley_core::field::Vec2,
-) {
-    if let Some(pending_press) = st.input.interaction_state.pending_core_press.clone() {
-        let dx = effective_sx - pending_press.press_global_sx;
-        let dy = effective_sy - pending_press.press_global_sy;
-        const CORE_CLICK_DRAG_THRESHOLD_PX: f32 = 8.0;
-        if dx.hypot(dy) >= CORE_CLICK_DRAG_THRESHOLD_PX {
-            st.input.interaction_state.pending_core_press = None;
-            if st.model.field.node(pending_press.node_id).is_some() {
-                begin_drag(
-                    st,
-                    ps,
-                    backend,
-                    HitNode {
-                        node_id: pending_press.node_id,
-                        on_titlebar: true,
-                        is_core: true,
-                    },
-                    ButtonFrame {
-                        ws_w: local_w,
-                        ws_h: local_h,
-                        global_sx: effective_sx,
-                        global_sy: effective_sy,
-                        sx: local_sx,
-                        sy: local_sy,
-                        world_now: pointer_world,
-                        workspace_active: false,
-                    },
-                    pointer_world,
-                    false,
-                );
-                backend.request_redraw();
-            }
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(super) fn handle_drag_motion(
-    st: &mut Halley,
-    backend: &impl BackendView,
-    mods: &ModState,
-    ps: &mut PointerState,
-    drag_mod_ok: bool,
-    target_monitor: &str,
-    local_w: i32,
-    local_h: i32,
-    local_sx: f32,
-    local_sy: f32,
-    pointer_world: halley_core::field::Vec2,
-    now: Instant,
-) -> bool {
-    let Some(drag) = ps.drag else {
-        st.input.interaction_state.cluster_join_candidate = None;
-        return false;
-    };
-
-    if ps.resize.is_some() || !drag_mod_ok {
-        let joined = !drag_mod_ok && st.commit_ready_cluster_join_for_node(drag.node_id, now);
-        st.set_drag_authority_node(None);
-        st.end_carry_state_tracking(drag.node_id);
-        ps.drag = None;
-        st.input.interaction_state.active_drag = None;
-        if joined {
-            backend.request_redraw();
-        }
-        return joined;
-    }
-
-    let mut next_drag = drag;
-    let drag_allow_monitor_transfer = super::pointer_frame::active_pointer_binding(st, mods, 0x110)
-        == Some(halley_config::PointerBindingAction::FieldJump);
-    next_drag.allow_monitor_transfer = drag_allow_monitor_transfer;
-    let dt = now
-        .saturating_duration_since(next_drag.last_update_at)
-        .as_secs_f32()
-        .max(1.0 / 240.0);
-    let raw_velocity = halley_core::field::Vec2 {
-        x: (pointer_world.x - next_drag.last_pointer_world.x) / dt,
-        y: (pointer_world.y - next_drag.last_pointer_world.y) / dt,
-    };
-    let max_drag_speed = 800.0f32;
-    let clamp_axis = |v: f32| v.clamp(-max_drag_speed, max_drag_speed);
-    next_drag.release_velocity = halley_core::field::Vec2 {
-        x: next_drag.release_velocity.x * 0.35 + clamp_axis(raw_velocity.x) * 0.65,
-        y: next_drag.release_velocity.y * 0.35 + clamp_axis(raw_velocity.y) * 0.65,
-    };
-    next_drag.last_pointer_world = pointer_world;
-    next_drag.last_update_at = now;
-    let desired_to = halley_core::field::Vec2 {
-        x: pointer_world.x - next_drag.current_offset.x,
-        y: pointer_world.y - next_drag.current_offset.y,
-    };
-
-    update_drag_edge_pan(
-        st,
-        drag.node_id,
-        target_monitor,
-        desired_to,
-        dt,
-        drag_allow_monitor_transfer,
-        &mut next_drag,
-    );
-
-    let should_center = st.runtime.tuning.center_window_to_mouse
-        && (!next_drag.center_latched
-            || next_drag.current_offset.x.abs() > f32::EPSILON
-            || next_drag.current_offset.y.abs() > f32::EPSILON);
-    if should_center {
-        next_drag.current_offset = halley_core::field::Vec2 { x: 0.0, y: 0.0 };
-        next_drag.center_latched = true;
-    }
-    st.input.interaction_state.drag_authority_velocity = next_drag.release_velocity;
-    st.input.interaction_state.active_drag = Some(ActiveDragState {
-        node_id: drag.node_id,
-        allow_monitor_transfer: drag_allow_monitor_transfer,
-        edge_pan_eligible: next_drag.edge_pan_eligible,
-        current_offset: next_drag.current_offset,
-        pointer_monitor: target_monitor.to_string(),
-        pointer_workspace_size: (local_w, local_h),
-        pointer_screen_local: (local_sx, local_sy),
-        edge_pan_x: next_drag.edge_pan_x,
-        edge_pan_y: next_drag.edge_pan_y,
-    });
-    ps.drag = Some(next_drag);
-    let _ = update_cluster_join_candidate(st, drag.node_id, target_monitor, desired_to, now);
-    backend.request_redraw();
-    true
-}
-
-fn update_drag_edge_pan(
-    st: &mut Halley,
-    node_id: halley_core::field::NodeId,
-    target_monitor: &str,
-    desired_to: halley_core::field::Vec2,
-    dt: f32,
-    drag_allow_monitor_transfer: bool,
-    next_drag: &mut crate::interaction::types::DragCtx,
-) {
-    if !drag_allow_monitor_transfer
-        && next_drag.edge_pan_eligible
-        && let Some(owner_monitor) = st
-            .model
-            .monitor_state
-            .node_monitor
-            .get(&node_id)
-            .cloned()
-            .or_else(|| Some(target_monitor.to_string()))
-    {
-        if let Some((clamped_center, edge_contact)) = st.dragged_node_edge_pan_clamp(
-            owner_monitor.as_str(),
-            node_id,
-            desired_to,
-            halley_core::field::Vec2 {
-                x: next_drag.edge_pan_x.sign(),
-                y: next_drag.edge_pan_y.sign(),
-            },
-        ) {
-            const EDGE_PAN_PRESSURE_THRESHOLD: f32 = 56.0;
-            const EDGE_PAN_PRESSURE_DECAY_PER_SEC: f32 = 44.0;
-            const EDGE_PAN_PRESSURE_BUILD_PER_SEC: f32 = 86.0;
-            const EDGE_PAN_PRESSURE_DEPTH_NORM: f32 = 18.0;
-            const EDGE_PAN_RELEASE_DISTANCE: f32 = 24.0;
-
-            next_drag.edge_pan_pressure.x =
-                (next_drag.edge_pan_pressure.x - EDGE_PAN_PRESSURE_DECAY_PER_SEC * dt).max(0.0);
-            next_drag.edge_pan_pressure.y =
-                (next_drag.edge_pan_pressure.y - EDGE_PAN_PRESSURE_DECAY_PER_SEC * dt).max(0.0);
-
-            if edge_contact.x < 0.0 {
-                let depth = (clamped_center.x - desired_to.x).max(0.0);
-                let build = (depth / EDGE_PAN_PRESSURE_DEPTH_NORM).clamp(0.0, 1.25);
-                next_drag.edge_pan_pressure.x += EDGE_PAN_PRESSURE_BUILD_PER_SEC * build * dt;
-            } else if edge_contact.x > 0.0 {
-                let depth = (desired_to.x - clamped_center.x).max(0.0);
-                let build = (depth / EDGE_PAN_PRESSURE_DEPTH_NORM).clamp(0.0, 1.25);
-                next_drag.edge_pan_pressure.x += EDGE_PAN_PRESSURE_BUILD_PER_SEC * build * dt;
-            } else {
-                next_drag.edge_pan_pressure.x = 0.0;
-            }
-
-            if edge_contact.y < 0.0 {
-                let depth = (clamped_center.y - desired_to.y).max(0.0);
-                let build = (depth / EDGE_PAN_PRESSURE_DEPTH_NORM).clamp(0.0, 1.25);
-                next_drag.edge_pan_pressure.y += EDGE_PAN_PRESSURE_BUILD_PER_SEC * build * dt;
-            } else if edge_contact.y > 0.0 {
-                let depth = (desired_to.y - clamped_center.y).max(0.0);
-                let build = (depth / EDGE_PAN_PRESSURE_DEPTH_NORM).clamp(0.0, 1.25);
-                next_drag.edge_pan_pressure.y += EDGE_PAN_PRESSURE_BUILD_PER_SEC * build * dt;
-            } else {
-                next_drag.edge_pan_pressure.y = 0.0;
-            }
-
-            next_drag.edge_pan_x = match next_drag.edge_pan_x {
-                DragAxisMode::Free => {
-                    if edge_contact.x < 0.0
-                        && next_drag.edge_pan_pressure.x >= EDGE_PAN_PRESSURE_THRESHOLD
-                    {
-                        DragAxisMode::EdgePanNeg
-                    } else if edge_contact.x > 0.0
-                        && next_drag.edge_pan_pressure.x >= EDGE_PAN_PRESSURE_THRESHOLD
-                    {
-                        DragAxisMode::EdgePanPos
-                    } else {
-                        DragAxisMode::Free
-                    }
-                }
-                DragAxisMode::EdgePanNeg => {
-                    if desired_to.x > clamped_center.x + EDGE_PAN_RELEASE_DISTANCE {
-                        next_drag.edge_pan_pressure.x = 0.0;
-                        DragAxisMode::Free
-                    } else {
-                        DragAxisMode::EdgePanNeg
-                    }
-                }
-                DragAxisMode::EdgePanPos => {
-                    if desired_to.x < clamped_center.x - EDGE_PAN_RELEASE_DISTANCE {
-                        next_drag.edge_pan_pressure.x = 0.0;
-                        DragAxisMode::Free
-                    } else {
-                        DragAxisMode::EdgePanPos
-                    }
-                }
-            };
-            next_drag.edge_pan_y = match next_drag.edge_pan_y {
-                DragAxisMode::Free => {
-                    if edge_contact.y < 0.0
-                        && next_drag.edge_pan_pressure.y >= EDGE_PAN_PRESSURE_THRESHOLD
-                    {
-                        DragAxisMode::EdgePanNeg
-                    } else if edge_contact.y > 0.0
-                        && next_drag.edge_pan_pressure.y >= EDGE_PAN_PRESSURE_THRESHOLD
-                    {
-                        DragAxisMode::EdgePanPos
-                    } else {
-                        DragAxisMode::Free
-                    }
-                }
-                DragAxisMode::EdgePanNeg => {
-                    if desired_to.y > clamped_center.y + EDGE_PAN_RELEASE_DISTANCE {
-                        next_drag.edge_pan_pressure.y = 0.0;
-                        DragAxisMode::Free
-                    } else {
-                        DragAxisMode::EdgePanNeg
-                    }
-                }
-                DragAxisMode::EdgePanPos => {
-                    if desired_to.y < clamped_center.y - EDGE_PAN_RELEASE_DISTANCE {
-                        next_drag.edge_pan_pressure.y = 0.0;
-                        DragAxisMode::Free
-                    } else {
-                        DragAxisMode::EdgePanPos
-                    }
-                }
-            };
-
-            let edge_pan_direction = halley_core::field::Vec2 {
-                x: next_drag.edge_pan_x.sign(),
-                y: next_drag.edge_pan_y.sign(),
-            };
-            let edge_pan_active = edge_pan_direction.x != 0.0 || edge_pan_direction.y != 0.0;
-            let indicator_direction = if edge_pan_active {
-                edge_pan_direction
-            } else {
-                edge_contact
-            };
-
-            st.input.interaction_state.grabbed_edge_pan_active = edge_pan_active;
-            st.input.interaction_state.grabbed_edge_pan_direction = indicator_direction;
-            st.input.interaction_state.grabbed_edge_pan_pressure = next_drag.edge_pan_pressure;
-            st.input.interaction_state.grabbed_edge_pan_monitor = ((indicator_direction.x != 0.0
-                || indicator_direction.y != 0.0)
-                && (next_drag.edge_pan_pressure.x > 0.0 || next_drag.edge_pan_pressure.y > 0.0))
-                .then(|| owner_monitor.clone());
-            return;
-        }
-    }
-
-    st.input.interaction_state.grabbed_edge_pan_active = false;
-    st.input.interaction_state.grabbed_edge_pan_direction =
-        halley_core::field::Vec2 { x: 0.0, y: 0.0 };
-    st.input.interaction_state.grabbed_edge_pan_pressure =
-        halley_core::field::Vec2 { x: 0.0, y: 0.0 };
-    st.input.interaction_state.grabbed_edge_pan_monitor = None;
-    next_drag.edge_pan_x = DragAxisMode::Free;
-    next_drag.edge_pan_y = DragAxisMode::Free;
-    next_drag.edge_pan_pressure = halley_core::field::Vec2 { x: 0.0, y: 0.0 };
-}
