+++ crates/halley-wl/src/lib.rs
@@ -2,6 +2,7 @@ pub mod activity;
 pub mod animation;
 pub mod bootstrap;
 pub(crate) mod backend;
+pub(crate) mod compositor;
 pub(crate) mod input;
 pub(crate) mod interaction;
 pub(crate) mod ipc;
@@ -9,8 +10,8 @@ pub(crate) mod overlay;
 pub mod render;
 pub(crate) mod spatial;
 pub mod state;
-pub(crate) mod surface_ops;
 pub(crate) mod wayland;
 pub(crate) mod wm;
 
 pub use bootstrap::{run, run_winit};
+
